requirements.txt
Django
psycopg2-binary
djangorestframework
django-webpack-loader